requirements.txt
mlflow==2.14.3
numpy==2.0.1
opencv_python==4.10.0.84
pandas==2.2.2
Pillow==9.4.0
Pillow==10.4.0
torch==2.1.0
torchvision==0.16.0
model==0.6.0
timm==1.0.8
torch==2.1.0
data==0.4
matplotlib==3.9.1
mlflow==2.14.3
numpy==2.0.1
Pillow==9.4.0
Pillow==10.4.0
ray==2.33.0
scikit_learn==1.5.1
seaborn==0.13.2
tools==0.1.9
torch==2.1.0
